refactor(backend): extract fan control backends into separate modules
