+++ backend/app/services/power/fan_control.py
@@ -2,17 +2,16 @@
 Fan control service for PWM fan management.
 
 Supports both Linux hardware (via hwmon sysfs) and development simulation.
+Backend implementations are in fan_backend_dev.py and fan_backend_linux.py.
 """
 import asyncio
 import json
 import logging
-import os
 import time
 from abc import ABC, abstractmethod
 from collections import deque
 from dataclasses import dataclass
 from datetime import datetime, timezone
-from pathlib import Path
 from typing import Dict, List, Optional, Tuple
 
 from sqlalchemy import select, desc
@@ -24,6 +23,10 @@ from app.schemas.fans import FanMode, FanCurvePoint
 
 logger = logging.getLogger(__name__)
 
+
+# ---------------------------------------------------------------------------
+# Shared data types
+# ---------------------------------------------------------------------------
 
 @dataclass
 class FanData:
@@ -60,6 +63,10 @@ class TempSensorData:
     is_cpu_sensor: bool
     current_temp: Optional[float]
 
+
+# ---------------------------------------------------------------------------
+# Backend ABC
+# ---------------------------------------------------------------------------
 
 class FanControlBackend(ABC):
     """Abstract base class for fan control backends."""
@@ -99,557 +106,14 @@ class FanControlBackend(ABC):
         pass
 
 
-class DevFanControlBackend(FanControlBackend):
-    """Development backend with simulated fans."""
+# Re-export backend classes for backward compatibility
+from app.services.power.fan_backend_dev import DevFanControlBackend  # noqa: E402
+from app.services.power.fan_backend_linux import LinuxFanControlBackend  # noqa: E402
 
-    def __init__(self, config: Settings):
-        self.config = config
-        self._fans: Dict[str, Dict] = {}
-        self._temps: Dict[str, float] = {}
-        self._initialize_simulated_fans()
 
-    def _initialize_simulated_fans(self):
-        """Initialize 3 simulated PWM fans."""
-        self._fans = {
-            "dev_cpu_fan": {
-                "name": "CPU Fan (Simulated)",
-                "pwm_percent": 50,
-                "target_rpm": 2000,
-                "current_rpm": 2000,
-                "min_rpm": 1000,
-                "max_rpm": 3000,
-                "temp_sensor_id": "dev_cpu_temp",
-                "last_update": time.time(),
-            },
-            "dev_case_fan_1": {
-                "name": "Case Fan 1 (Simulated)",
-                "pwm_percent": 40,
-                "target_rpm": 1200,
-                "current_rpm": 1200,
-                "min_rpm": 800,
-                "max_rpm": 2000,
-                "temp_sensor_id": "dev_package_temp",
-                "last_update": time.time(),
-            },
-            "dev_case_fan_2": {
-                "name": "Case Fan 2 (Simulated)",
-                "pwm_percent": 40,
-                "target_rpm": 1200,
-                "current_rpm": 1200,
-                "min_rpm": 800,
-                "max_rpm": 2000,
-                "temp_sensor_id": "dev_ambient_temp",
-                "last_update": time.time(),
-            },
-        }
-
-        # Initialize simulated temperatures
-        self._temps = {
-            "dev_cpu_temp": 45.0,
-            "dev_package_temp": 42.0,
-            "dev_ambient_temp": 35.0,
-        }
-
-    async def is_available(self) -> bool:
-        """Dev backend is always available."""
-        return True
-
-    async def get_fans(self) -> List[FanData]:
-        """Get simulated fans."""
-        self._update_simulated_state()
-
-        fans = []
-        for fan_id, fan_data in self._fans.items():
-            # Get default curve if not configured
-            curve_points = self._get_default_curve()
-
-            fans.append(FanData(
-                fan_id=fan_id,
-                name=fan_data["name"],
-                rpm=fan_data["current_rpm"],
-                pwm_percent=fan_data["pwm_percent"],
-                temperature_celsius=self._temps.get(fan_data["temp_sensor_id"]),
-                mode=FanMode.AUTO,  # Default mode
-                min_pwm_percent=self.config.fan_min_pwm_percent,
-                max_pwm_percent=100,
-                emergency_temp_celsius=self.config.fan_emergency_temp_celsius,
-                temp_sensor_id=fan_data["temp_sensor_id"],
-                curve_points=curve_points,
-                is_active=True,
-            ))
-
-        return fans
-
-    async def set_pwm(self, fan_id: str, pwm_percent: int) -> bool:
-        """Set simulated PWM value."""
-        if fan_id not in self._fans:
-            logger.warning(f"Fan {fan_id} not found")
-            return False
-
-        pwm_percent = max(0, min(100, pwm_percent))
-        fan = self._fans[fan_id]
-        fan["pwm_percent"] = pwm_percent
-
-        # Calculate target RPM
-        pwm_ratio = pwm_percent / 100.0
-        fan["target_rpm"] = int(fan["min_rpm"] + (fan["max_rpm"] - fan["min_rpm"]) * pwm_ratio)
-        fan["last_update"] = time.time()
-
-        logger.debug(f"Set {fan_id} PWM to {pwm_percent}% (target RPM: {fan['target_rpm']})")
-        return True
-
-    async def get_temperature(self, sensor_id: str) -> Optional[float]:
-        """Get simulated temperature."""
-        return self._temps.get(sensor_id)
-
-    async def get_available_temp_sensors(self) -> List[TempSensorData]:
-        """List simulated temperature sensors."""
-        self._update_simulated_state()
-        return [
-            TempSensorData(
-                sensor_id="dev_cpu_temp",
-                device_name="k10temp (Simulated)",
-                label="Tctl",
-                is_cpu_sensor=True,
-                current_temp=self._temps.get("dev_cpu_temp"),
-            ),
-            TempSensorData(
-                sensor_id="dev_package_temp",
-                device_name="k10temp (Simulated)",
-                label="Tccd1",
-                is_cpu_sensor=True,
-                current_temp=self._temps.get("dev_package_temp"),
-            ),
-            TempSensorData(
-                sensor_id="dev_ambient_temp",
-                device_name="it8688e (Simulated)",
-                label="Board",
-                is_cpu_sensor=False,
-                current_temp=self._temps.get("dev_ambient_temp"),
-            ),
-        ]
-
-    def _update_simulated_state(self):
-        """Update simulated fan RPM and temperatures with realistic behavior."""
-        current_time = time.time()
-
-        # Simulate temperature fluctuations based on CPU usage
-        import random
-        base_cpu_temp = 45.0 + random.uniform(-3, 5)
-        self._temps["dev_cpu_temp"] = base_cpu_temp
-        self._temps["dev_package_temp"] = base_cpu_temp - 3 + random.uniform(-2, 2)
-        self._temps["dev_ambient_temp"] = 35.0 + random.uniform(-1, 2)
-
-        # Update fan RPM with latency and fluctuation
-        for fan_id, fan_data in self._fans.items():
-            elapsed = current_time - fan_data["last_update"]
-
-            # Gradual RPM transition (2-3 second lag)
-            if fan_data["current_rpm"] != fan_data["target_rpm"]:
-                rpm_diff = fan_data["target_rpm"] - fan_data["current_rpm"]
-                rpm_change = rpm_diff * min(elapsed / 2.5, 1.0)  # 2.5s to reach target
-                fan_data["current_rpm"] = int(fan_data["current_rpm"] + rpm_change)
-
-            # Add realistic RPM fluctuation (±30 RPM)
-            fluctuation = random.randint(-30, 30)
-            fan_data["current_rpm"] = max(
-                fan_data["min_rpm"],
-                min(fan_data["max_rpm"], fan_data["current_rpm"] + fluctuation)
-            )
-
-    def _get_default_curve(self) -> List[FanCurvePoint]:
-        """Get default temperature-PWM curve."""
-        return [
-            FanCurvePoint(temp=35, pwm=30),
-            FanCurvePoint(temp=50, pwm=50),
-            FanCurvePoint(temp=70, pwm=80),
-            FanCurvePoint(temp=85, pwm=100),
-        ]
-
-
-class LinuxFanControlBackend(FanControlBackend):
-    """Linux hardware backend using hwmon sysfs."""
-
-    def __init__(self, config: Settings):
-        self.config = config
-        self._hwmon_base = Path("/sys/class/hwmon")
-        self._fan_cache: Dict[str, Dict] = {}
-        self._has_write_permission = False
-
-    async def is_available(self) -> bool:
-        """Check if hwmon is available."""
-        if not self._hwmon_base.exists():
-            logger.info("hwmon not available (not on Linux or no sensors)")
-            return False
-
-        # Scan for PWM fans
-        fans = await self._scan_pwm_fans()
-        if not fans:
-            logger.info("No PWM fans found in hwmon")
-            return False
-
-        logger.info(f"Found {len(fans)} PWM fan(s) in hwmon")
-        await self._check_write_permission()
-        return True
-
-    async def _check_write_permission(self) -> None:
-        """Proactively test write permission on first PWM file."""
-        if not self._fan_cache:
-            return
-
-        first_fan = next(iter(self._fan_cache.values()))
-        pwm_path = first_fan["pwm_path"]
-
-        # Fast check: direct write permission
-        if os.access(pwm_path, os.W_OK):
-            self._has_write_permission = True
-            logger.info("Fan control: direct write permission available")
-            return
-
-        # Fallback: test sudo tee (read current value, write it back unchanged)
-        current = await self._read_hwmon_file(pwm_path)
-        if current is not None:
-            try:
-                import subprocess
-                result = subprocess.run(
-                    ["sudo", "-n", "tee", str(pwm_path)],
-                    input=str(current).encode(),
-                    capture_output=True,
-                    timeout=5,
-                )
-                if result.returncode == 0:
-                    self._has_write_permission = True
-                    logger.info("Fan control: write permission available via sudo tee")
-                    return
-            except Exception:
-                pass
-
-        logger.info("Fan control: no write permission (readonly mode)")
-
-    async def get_fans(self) -> List[FanData]:
-        """Get hardware fans from hwmon (uses cache from startup scan)."""
-
-        fans = []
-        for fan_id, fan_info in self._fan_cache.items():
-            # Read current state
-            pwm_value = await self._read_hwmon_file(fan_info["pwm_path"])
-            pwm_percent = self._pwm_to_percent(pwm_value) if pwm_value is not None else 0
-
-            rpm_value = await self._read_hwmon_file(fan_info["fan_input_path"])
-            rpm = int(rpm_value) if rpm_value is not None else None
-
-            temp_celsius = None
-            if fan_info.get("temp_path"):
-                temp_value = await self._read_hwmon_file(fan_info["temp_path"])
-                if temp_value is not None:
-                    temp_celsius = float(temp_value) / 1000.0  # millidegrees to degrees
-
-            fans.append(FanData(
-                fan_id=fan_id,
-                name=fan_info["name"],
-                rpm=rpm,
-                pwm_percent=pwm_percent,
-                temperature_celsius=temp_celsius,
-                mode=FanMode.AUTO,  # Will be overridden by service from DB
-                min_pwm_percent=self.config.fan_min_pwm_percent,
-                max_pwm_percent=100,
-                emergency_temp_celsius=self.config.fan_emergency_temp_celsius,
-                temp_sensor_id=fan_info.get("temp_sensor_id"),
-                curve_points=self._get_default_curve(),
-                is_active=True,
-            ))
-
-        return fans
-
-    async def set_pwm(self, fan_id: str, pwm_percent: int) -> bool:
-        """Set hardware PWM value."""
-        if fan_id not in self._fan_cache:
-            logger.warning(f"Fan {fan_id} not found in cache")
-            return False
-
-        fan_info = self._fan_cache[fan_id]
-        pwm_path = fan_info["pwm_path"]
-        pwm_enable_path = fan_info.get("pwm_enable_path")
-
-        # Clamp to valid range
-        pwm_percent = max(0, min(100, pwm_percent))
-        pwm_value = self._percent_to_pwm(pwm_percent)
-
-        # First, set PWM enable to manual mode (1)
-        if pwm_enable_path:
-            success = await self._write_hwmon_file(pwm_enable_path, "1")
-            if not success:
-                logger.warning(f"Failed to set PWM enable for {fan_id}")
-
-        # Write PWM value
-        success = await self._write_hwmon_file(pwm_path, str(pwm_value))
-        if success:
-            logger.debug(f"Set {fan_id} PWM to {pwm_percent}% ({pwm_value}/255)")
-            return True
-        else:
-            logger.error(f"Failed to write PWM for {fan_id}")
-            return False
-
-    # CPU temperature driver names (same keywords as hardware/sensors.py)
-    _CPU_SENSOR_DRIVERS = {"k10temp", "coretemp", "cpu_thermal", "acpi"}
-
-    async def get_temperature(self, sensor_id: str) -> Optional[float]:
-        """Get temperature from hwmon sensor."""
-        if not sensor_id:
-            return None
-
-        # sensor_id format: "hwmon0_temp1"
-        parts = sensor_id.split("_")
-        if len(parts) != 2:
-            return None
-
-        hwmon_name, temp_name = parts
-        temp_path = self._hwmon_base / hwmon_name / f"{temp_name}_input"
-
-        temp_value = await self._read_hwmon_file(temp_path)
-        if temp_value is not None:
-            return float(temp_value) / 1000.0
-
-        return None
-
-    def _find_cpu_temp_sensor(self) -> Optional[Tuple[str, Path]]:
-        """Find CPU temperature sensor across all hwmon directories.
-
-        Searches for known CPU temperature drivers (k10temp, coretemp, etc.)
-        in any hwmon directory, not just the one containing the PWM fan.
-
-        Returns:
-            Tuple of (sensor_id, temp_path) or None if not found
-        """
-        if not self._hwmon_base.exists():
-            return None
-
-        for hwmon_dir in sorted(self._hwmon_base.iterdir()):
-            if not hwmon_dir.is_dir() or not hwmon_dir.name.startswith("hwmon"):
-                continue
-
-            name_file = hwmon_dir / "name"
-            if not name_file.exists():
-                continue
-
-            try:
-                driver_name = name_file.read_text().strip()
-            except Exception:
-                continue
-
-            if driver_name not in self._CPU_SENSOR_DRIVERS:
-                continue
-
-            # Found a CPU sensor driver — use its first temp input
-            for temp_file in sorted(hwmon_dir.glob("temp*_input")):
-                temp_num = temp_file.name.replace("temp", "").replace("_input", "")
-                sensor_id = f"{hwmon_dir.name}_temp{temp_num}"
-                logger.info(
-                    f"Found CPU temp sensor: {sensor_id} (driver={driver_name})"
-                )
-                return sensor_id, temp_file
-
-        return None
-
-    async def get_available_temp_sensors(self) -> List[TempSensorData]:
-        """List all available temperature sensors across all hwmon directories."""
-        sensors: List[TempSensorData] = []
-
-        if not self._hwmon_base.exists():
-            return sensors
-
-        for hwmon_dir in sorted(self._hwmon_base.iterdir()):
-            if not hwmon_dir.is_dir() or not hwmon_dir.name.startswith("hwmon"):
-                continue
-
-            name_file = hwmon_dir / "name"
-            device_name = "Unknown"
-            if name_file.exists():
-                try:
-                    device_name = name_file.read_text().strip()
-                except Exception:
-                    pass
-
-            is_cpu = device_name in self._CPU_SENSOR_DRIVERS
-
-            for temp_file in sorted(hwmon_dir.glob("temp*_input")):
-                temp_num = temp_file.name.replace("temp", "").replace("_input", "")
-                sensor_id = f"{hwmon_dir.name}_temp{temp_num}"
-
-                # Try to read label
-                label = None
-                label_file = hwmon_dir / f"temp{temp_num}_label"
-                if label_file.exists():
-                    try:
-                        label = label_file.read_text().strip()
-                    except Exception:
-                        pass
-
-                # Read current temperature
-                current_temp = None
-                temp_value = await self._read_hwmon_file(temp_file)
-                if temp_value is not None:
-                    current_temp = float(temp_value) / 1000.0
-
-                sensors.append(TempSensorData(
-                    sensor_id=sensor_id,
-                    device_name=device_name,
-                    label=label,
-                    is_cpu_sensor=is_cpu,
-                    current_temp=current_temp,
-                ))
-
-        return sensors
-
-    async def _scan_pwm_fans(self) -> Dict[str, Dict]:
-        """Scan hwmon for PWM fans.
-
-        Builds a new dict from sysfs. Only replaces the cache if the scan
-        found at least one fan (or the cache was empty), preventing
-        transient sysfs I/O failures from clearing known fans.
-
-        Prefers CPU temperature sensor (k10temp, coretemp) over local
-        board sensors for fan control.
-        """
-        new_cache: Dict[str, Dict] = {}
-
-        if not self._hwmon_base.exists():
-            if not self._fan_cache:
-                return {}
-            logger.debug("hwmon base missing but cache exists, keeping cached fans")
-            return self._fan_cache
-
-        # Find CPU temp sensor across all hwmon directories
-        cpu_sensor = self._find_cpu_temp_sensor()
-        cpu_sensor_id = cpu_sensor[0] if cpu_sensor else None
-        cpu_temp_path = cpu_sensor[1] if cpu_sensor else None
-
-        for hwmon_dir in self._hwmon_base.iterdir():
-            if not hwmon_dir.is_dir() or not hwmon_dir.name.startswith("hwmon"):
-                continue
-
-            # Read hwmon name
-            name_file = hwmon_dir / "name"
-            hwmon_name_value = "Unknown"
-            if name_file.exists():
-                try:
-                    hwmon_name_value = name_file.read_text().strip()
-                except Exception:
-                    pass
-
-            # Find PWM files
-            for pwm_file in hwmon_dir.glob("pwm[0-9]*"):
-                if "_" in pwm_file.name:  # Skip pwm1_enable, etc
-                    continue
-
-                pwm_num = pwm_file.name.replace("pwm", "")
-                fan_id = f"{hwmon_dir.name}_pwm{pwm_num}"
-
-                # Find corresponding fan input
-                fan_input_path = hwmon_dir / f"fan{pwm_num}_input"
-                if not fan_input_path.exists():
-                    continue
-
-                # Find PWM enable
-                pwm_enable_path = hwmon_dir / f"pwm{pwm_num}_enable"
-
-                # Prefer CPU sensor over local board sensor
-                if cpu_sensor_id:
-                    temp_sensor_id = cpu_sensor_id
-                    temp_path = cpu_temp_path
-                else:
-                    # Fallback: use first temp sensor in same hwmon dir
-                    temp_sensor_id = None
-                    temp_path = None
-                    for temp_file in hwmon_dir.glob("temp*_input"):
-                        temp_num = temp_file.name.replace("temp", "").replace("_input", "")
-                        temp_sensor_id = f"{hwmon_dir.name}_temp{temp_num}"
-                        temp_path = temp_file
-                        break
-
-                new_cache[fan_id] = {
-                    "name": f"{hwmon_name_value} PWM{pwm_num}",
-                    "pwm_path": pwm_file,
-                    "pwm_enable_path": pwm_enable_path if pwm_enable_path.exists() else None,
-                    "fan_input_path": fan_input_path,
-                    "temp_path": temp_path,
-                    "temp_sensor_id": temp_sensor_id,
-                }
-
-        if new_cache or not self._fan_cache:
-            self._fan_cache = new_cache
-            logger.info(f"Scanned hwmon: found {len(self._fan_cache)} PWM fan(s)")
-        else:
-            logger.warning(
-                f"hwmon scan found 0 fans but cache has {len(self._fan_cache)}, keeping cached fans"
-            )
-
-        return self._fan_cache
-
-    async def _read_hwmon_file(self, path: Path) -> Optional[int]:
-        """Read integer value from hwmon sysfs file."""
-        if not path or not path.exists():
-            return None
-
-        try:
-            value = path.read_text().strip()
-            return int(value)
-        except Exception as e:
-            logger.debug(f"Failed to read {path}: {e}")
-            return None
-
-    async def _write_hwmon_file(self, path: Path, value: str) -> bool:
-        """Write value to hwmon sysfs file."""
-        if not path or not path.exists():
-            return False
-
-        try:
-            # Try direct write first
-            path.write_text(value + "\n")
-            self._has_write_permission = True
-            return True
-        except PermissionError:
-            # Try with sudo tee fallback
-            try:
-                import subprocess
-                cmd = ["sudo", "tee", str(path)]
-                result = subprocess.run(
-                    cmd,
-                    input=value.encode(),
-                    capture_output=True,
-                    timeout=5
-                )
-                if result.returncode == 0:
-                    self._has_write_permission = True
-                    logger.debug(f"Wrote to {path} via sudo tee")
-                    return True
-                else:
-                    logger.warning(f"sudo tee failed for {path}: {result.stderr.decode()}")
-                    return False
-            except Exception as e:
-                logger.error(f"Failed to write {path} with sudo: {e}")
-                return False
-        except Exception as e:
-            logger.error(f"Failed to write {path}: {e}")
-            return False
-
-    def _pwm_to_percent(self, pwm_value: int) -> int:
-        """Convert PWM value (0-255) to percentage (0-100)."""
-        return round(pwm_value * 100 / 255)
-
-    def _percent_to_pwm(self, percent: int) -> int:
-        """Convert percentage (0-100) to PWM value (0-255)."""
-        return round(percent * 255 / 100)
-
-    def _get_default_curve(self) -> List[FanCurvePoint]:
-        """Get default temperature-PWM curve."""
-        return [
-            FanCurvePoint(temp=35, pwm=30),
-            FanCurvePoint(temp=50, pwm=50),
-            FanCurvePoint(temp=70, pwm=80),
-            FanCurvePoint(temp=85, pwm=100),
-        ]
-
+# ---------------------------------------------------------------------------
+# Service
+# ---------------------------------------------------------------------------
 
 class FanControlService:
     """Singleton service for managing fan control."""
